[ServerManage] Split into separate files
- Add exceptions.py for all exceptions.
- Add meta.py to be able to split commands from shared code.
- Add commands.py to hold only the commands and make it an abstract base
  class.
- No functional changes, only line moves and creating a metaclass for
  organizational purposes.
- Inspiration for metaclass from PhaseCoreX's PCXCogs, thanks!
