+++ cogs/servermanage/exceptions.py
@@ -1,0 +1,10 @@
+class InvalidAttachmentsError(Exception):
+    pass
+
+
+class InvalidFileError(Exception):
+    pass
+
+
+class InvalidImageError(Exception):
+    pass
