+++ cogs/servermanage/meta.py
@@ -1,0 +1,6 @@
+from abc import ABC
+from redbot.core import commands
+
+
+class ServerManageMeta(type(commands.Cog), type(ABC)):
+    pass
